backend-dev/requirements.txt
Django>=2.2.6
djangorestframework>=3.9.4
pytz>=2019.3
sqlparse>=0.3.0
uwsgi>=2.0
dj-database-url>=0.5
psycopg2>=2.8.4 --no-binary psycopg2
python-dateutil>=2.8.1
channels==2.4.0
channels-redis==2.4.2
pytest==5.4.1
pytest-asyncio==0.10.0
pytest-django==3.8.0
Pillow==7.1.1
django-cors-headers
